// Web/src/app/models/AddressVM.ts
export class AddressVM {
    ID_address: number;
    address_1: string;
    address_2: string;
    city: string;
    state: string;
    postal_code: string;
    country: string;
    company: number;
}

export class SearchAddressVM {
    TotalRows: number;
    ID_address: number;
    address_1: string;
    address_2: string;
    city: string;
    state: string;
    postal_code: string;
    country: string;
    company: number;
}